fix flakiness in warm eni tests
