+++ test/integration/ipamd/warm_target_test_PD_enabled.go
@@ -17,19 +17,12 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/aws/amazon-vpc-cni-k8s/test/framework"
-	"github.com/aws/aws-sdk-go/service/ec2"
-
 	k8sUtils "github.com/aws/amazon-vpc-cni-k8s/test/framework/resources/k8s/utils"
 	"github.com/aws/amazon-vpc-cni-k8s/test/framework/utils"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-var primaryInstance *ec2.Instance
-var f *framework.Framework
-var err error
 
 // IMPORTANT: THE NODEGROUP TO RUN THE TEST MUST NOT HAVE ANY POD
 // Ideally we should drain the node, but drain from go client is non trivial
@@ -113,7 +106,7 @@ var _ = Describe("test warm target variables", func() {
 				minIPTarget = 16
 			})
 
-			It("should have 1 prefixe", func() {})
+			It("should have 1 prefix", func() {})
 		})
 
 		Context("when MINIMUM_IP_TARGET = 6 and WARM_IP_TARGET = 10", func() {
@@ -258,7 +251,7 @@ var _ = Describe("test warm target variables", func() {
 				warmPrefixTarget = 0
 			})
 
-			It("should have 1 prefixe", func() {})
+			It("should have 1 prefix", func() {})
 		})
 
 		Context("when MINIMUM_IP_TARGET = 6, WARM_IP_TARGET = 10 and WARM_PREFIX_TARGET = 1", func() {
@@ -272,28 +265,3 @@ var _ = Describe("test warm target variables", func() {
 		})
 	})
 })
-
-func ceil(x, y int) int {
-	return (x + y - 1) / y
-}
-
-func Max(x, y int) int {
-	if x < y {
-		return y
-	}
-	return x
-}
-
-// MinIgnoreZero returns smaller of two number, if any number is zero returns the other number
-func MinIgnoreZero(x, y int) int {
-	if x == 0 {
-		return y
-	}
-	if y == 0 {
-		return x
-	}
-	if x < y {
-		return x
-	}
-	return y
-}
